Lab 1 : Warming up to Java
This lab objective is to create a program that reads in the user’s shoe size (no half sizes) and birth year. It then, performs arithmetic operations to get your age. This is a common mathematics riddle used to "magically" guess someone's likely shoe size and can be used without the info due to the order of operations.